+++ infomex/src/main/webapp/mapa.xhtml
@@ -9,8 +9,32 @@
         <title>Infomex</title>
     </h:head>
     <h:body>
-        <h1>Mapa chido</h1> <br/>
-
+        <h1>
+            <p:commandButton icon="ui-icon-grip-solid-horizontal" oncomplete="PF('sidebar1').show()" />
+        <center>Infomex</center>
+        </h1>
+        <p:sidebar widgetVar="sidebar1" baseZIndex="10000">
+            <h1 style="font-weight:normal">Menu</h1>
+            <p:commandButton value="Pagina principal" oncomplete="redirect('mapa.xhtml')"/>
+             <br/>
+             <p:commandButton value="Iniciar sesion" oncomplete="redirect()"/>
+             <br/>
+             <p:commandButton value="Registro" oncomplete="redirect('registro.xhtml')"/>
+             <br/>
+             <p:commandButton value="Agregar temas" oncomplete="redirect()"/>
+             <br/>
+             <p:commandButton value="Eliminar temas" oncomplete="redirect()"/>
+             <br/>
+            <p:commandButton value="Cancelar" oncomplete="PF('sidebar1').hide()"/>
+        </p:sidebar>
+        
+        <script> 
+            type="text/javascript"
+                redirect = function(ruta) {
+                    window.location.href = ruta;
+                }
+        </script>
+        
         <style>
 
         #myForm { }
@@ -65,19 +89,19 @@
 
 
 <h:form id="myForm">
-<p:gmap id="gmap" center="36.890257,30.707417" zoom="13" type="HYBRID" style="width:100%;height:550px"
-    model="#{markersView.simpleModel}" onPointClick="handlePointClick(event);" widgetVar="map" >
+<p:gmap id="gmap" center="19.432590, -99.131205" zoom="13" type="HYBRID" style="width:100%;height:550px"
+        model="#{administradorMarcador.simpleModel}" onPointClick="handlePointClick(event);" widgetVar="map" >
     
-    <p:ajax event="overlaySelect" listener="#{markersView.onMarkerSelect}" />
+    <p:ajax event="overlaySelect" listener="#{administradorMarcador.onMarkerSelect}" />
     
         <p:gmapInfoWindow id="infoWindow">
             <p:outputPanel style="text-align: center; display: block; margin: auto">
  
-                    <h:outputText value= "#{markersView.marker.title}" style="font-weight:bold"/>
+                    <h:outputText value= "#{administradorMarcador.marker.title}" style="font-weight:bold"/>
                     <br />
-                    <h:outputText value= "#{markersView.marker.getData().getDescripcion_mar()}" />
+                    <h:outputText value= "#{administradorMarcador.marker.getData().getDescripcion_mar()}" />
                     <br />
-                    <p:commandButton value="Eliminar" icon="pi pi-search" action="#{markersView.elimina}" update=":messages" oncomplete="window.location.reload()"/>
+                    <p:commandButton value="Eliminar" icon="pi pi-search" action="#{administradorMarcador.eliminaMarcador}" update=":messages" oncomplete="window.location.reload()"/>
                     
             </p:outputPanel>
         </p:gmapInfoWindow>
@@ -88,17 +112,17 @@
     <h:form prependId="false">
         <h:panelGrid columns="2">
             <h:outputLabel for="title" value="Nombre:" />
-            <p:inputText id="title" value="#{markersView.nombre_mar}" required="true"/>
+            <p:inputText id="title" value="#{administradorMarcador.nombre_mar}" required="true"/>
             <h:outputLabel for="descripcion" value="Descripcion:" />
-            <p:inputTextarea id="descripcion" value="#{markersView.descripcion_mar}" required="true"/>
+            <p:inputTextarea id="descripcion" value="#{administradorMarcador.descripcion_mar}" required="true"/>
             <f:facet name="footer">
-                <p:commandButton value="Add" action="#{markersView.addMarker}" update=":messages" oncomplete="markerAddComplete()"/>
+                <p:commandButton value="Add" action="#{administradorMarcador.agregaMarcador}" update=":messages" oncomplete="markerAddComplete()"/>
                 <p:commandButton value="Cancel" onclick="return cancel()" />
             </f:facet>
         </h:panelGrid>
  
-        <h:inputHidden id="lat" value="#{markersView.latitud}" />
-        <h:inputHidden id="lng" value="#{markersView.longitud}" />
+        <h:inputHidden id="lat" value="#{administradorMarcador.latitud}" />
+        <h:inputHidden id="lng" value="#{administradorMarcador.longitud}" />
     </h:form>
 </p:dialog>
