+++ infomex/src/main/webapp/menu.html
@@ -1,0 +1,80 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <title>Refugio Huellitas</title>
+
+
+    <link href="http://getbootstrap.com/dist/css/bootstrap.min.css" rel="stylesheet">
+
+    <link href="http://getbootstrap.com/examples/jumbotron-narrow/jumbotron-narrow.css" rel="stylesheet">
+<meta charset="utf-8">
+
+</head>
+
+<body>
+  <body background="imagenes/fondoPerro.jpg">
+    <header style="background: rgba(75,100,104,.9)">
+      <form class="navbar-form navbar-left" role="search">
+        <div class="container">
+          <div class="row">
+            <div class="col-lg-1">
+              <button type="submit" class="btn btn-primary">Entrar</button>
+            </div>
+            <div class="col-lg-3">
+              <input class="form-control" placeholder="Name" type="text">
+            </div>
+            <div class="col-lg-3">
+              <input class="form-control" placeholder="Password" type="text">
+            </div>
+          </div>
+        </div>
+      </form>
+    </header>
+    <DIV STYLE="position:absolute; top:25px; left:900px; visibility:visible z-index:1">
+    <IMG height="200" SRC="imagenes/logo.png" width="250">
+    </div>
+
+    <div class="container">
+        <div class="jumbotron" style="background-color: rgba(255, 255, 255, 0.5)">
+            <h1>Refugio Huellitas</h1>
+            <p class="lead"></p>
+            <p><a class="btn btn-lg btn-success" href="showSignUp" role="button">Registrate ahora!</a><br>
+            </p>
+        </div>
+
+        <div class="row marketing">
+          <div class="col-lg-4">
+            <h4>Citas</h4>
+            <p>Es momento de conseguirle a tu fiel amigo una pareja a su nivel...y a su gusto</p>
+          </div>
+
+          <div class="col-lg-4">
+            <h4>Mision</h4>
+            <p>El beneficio de mascotas y dueños para poderse relacionar de una manera mas eficaz</p>
+          </div>
+
+
+                    <div class="col-lg-4">
+                      <h4>Comienza ahora</h4>
+                      <p>Unete a esta gran comunidad dedicada a la reproduccion responsable de nuestras mascotass</p>
+                    </div>
+
+          <!-- <div class="col-lg-6">
+            {% for item in items %}
+              <h4>{{ item.name }}</h4>
+              <p>{{ item.age }}</p>
+            {% endfor %}
+          </div> -->
+
+
+        </div>
+
+        <footer class="footer">
+            <p>&copy; Company 2015</p>
+        </footer>
+
+    </div>
+</body>
+
+</html>
